refactor: Make Channel nullable and adjust null check on Channel in DiscordService and MonitorWorker

+++ Workers/MonitorWorker.cs
@@ -80,10 +80,11 @@ public class MonitorWorker : BackgroundService
 
             foreach (var video in videos)
             {
-                if (!video.Channel.Monitoring)
+                // Channel exists and is not monitoring
+                if (null != video.Channel && !video.Channel.Monitoring)
                 {
                     videoService.DeleteVideo(video);
-                    _logger.LogInformation("Remove scheduled video {videoId} because channel {channelId} is not monitoring", video.id, video.Channel.id);
+                    _logger.LogInformation("Remove scheduled video {videoId} because channel {channelId} is not monitoring.", video.id, video.Channel.id);
                 }
                 else
                 {
